feat: production transformation — Tasks 1-10 complete
Security: PyJWT + Argon2id, rate limiting, secret_key env validation
Database: Float→Numeric(15,2) everywhere, Alembic migration, composite indexes
Analytics: all aggregations rewritten as SQL GROUP BY / LEFT JOIN / HAVING
AI copilot: Claude Haiku with tool calling, agentic loop, ChatHistory persistence
Pagination: Page[T] generic envelope on /transactions and /budgets
Frontend: TanStack Query v5 (11 hooks), React.lazy for all 10 pages
Rewrites: BudgetsPage, DebtsPage, InvestmentsPage, SalaryPage → real API endpoints
Docker: multi-stage Dockerfiles (backend + frontend), docker-compose dev+prod
CI: GitHub Actions — ruff + pytest, tsc + build, docker build with GHA cache
Tests: 60 pytest tests (auth, analytics SQL, CSV edge cases), 60/60 pass

Bug fix: ChatHistoryResponse missing import in finance.py (NameError in prod)

+++ finflow-ai/backend/app/schemas/user.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import EmailStr
 
 from app.schemas.common import TimestampedResponse
@@ -7,5 +9,5 @@ class UserResponse(TimestampedResponse):
     id: int
     email: EmailStr
     full_name: str
-    monthly_income: float
+    monthly_income: Decimal
     currency: str
